% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file}
\alias{check_file}
\title{Check file type and structure is correct}
\usage{
check_file(data)
}
\arguments{
\item{data}{file to check for problems}
}
\value{
NULL or a character vector#'
}
\description{
Check file type and structure is correct
}
\author{
Hugh Jones \email{hj.hujones.hugh@googlemail.com}
}
