% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_atoms.R
\name{count_atoms}
\alias{count_atoms}
\title{Count atoms in a formula}
\usage{
count_atoms(formula, element)
}
\arguments{
\item{formula}{A chemical formula}

\item{element}{A chemical element}
}
\value{
an integer
}
\description{
Count atoms in a formula
}
\examples{
count_atoms("C6H12O6","C")
}
\author{
Hugh Jones \email{hj.hujones.hugh@googlemail.com}
}
