% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{detected_data}
\alias{detected_data}
\title{Molecular formula assignments found at least three datasets}
\format{
A list with 8 levels
\describe{
\item{ESFA_neg}{Data table holding negative mode assignments for ESFA}
\item{ESFA_pos}{Data table holding positive mode assignments for ESFA}
\item{PLFA_neg}{Data table holding negative mode assignments for PLFA}
\item{PLFA_pos}{Data table holding positive mode assignments for PLFA}
\item{SRFA_neg}{Data table holding negative mode assignments for SRFA}
\item{SRFA_pos}{Data table holding positive mode assignments for SRFA}
\item{SRNOM_neg}{Data table holding negative mode assignments for SRNOM}
\item{SRNOM_pos}{Data table holding positive mode assignments for SRNOM}
}
Each data frame has 12 variables:
\describe{
\item{C}{Number of carbon atoms in the formula}
\item{H}{Number of hydrogen atoms in the formula}
\item{O}{Number of oxygen atoms in the formula}
\item{N}{Number of nitrogen atoms in the formula}
\item{S}{Number of sulfur atoms in the formula}
\item{Na}{Number of sodium atoms in the formula}
\item{mz}{theoretical calculated m/z of the formula}
\item{formula}{assigned formula}
\item{H_C}{H/C ratio}
\item{O_C}{O/C ratio}
\item{dbe}{Double Bond Equivalent}
}
}
\usage{
detected_data
}
\description{
A list of 8 data frames
}
\seealso{
Other study data can be found in \code{common_data} and \code{sample_metrics}
}
\keyword{datasets}
