% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_palette.R
\name{expand_palette}
\alias{expand_palette}
\title{Expand colour palette}
\usage{
expand_palette(vector, default_colour = "black", named_colours)
}
\arguments{
\item{vector}{vector to generate colours for}

\item{default_colour}{default colour to use}

\item{named_colours}{named vector for colours to use}
}
\value{
an vector of colours
}
\description{
Expand colour palette
}
\examples{
expand_palette(c("a","b","c"), "black", c("c" = "red"))
}
\author{
Hugh Jones \email{hj.hujones.hugh@googlemail.com}
}
