% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Help.R
\name{mod_Help_ui}
\alias{mod_Help_ui}
\alias{mod_Help_server}
\title{mod_Help_ui and mod_mod_Help_server}
\usage{
mod_Help_ui(id)

mod_Help_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
