% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputData.R
\name{mod_inputData_ui}
\alias{mod_inputData_ui}
\alias{mod_inputData_server}
\title{mod_inputData_ui and mod_inputData_server}
\usage{
mod_inputData_ui(id)

mod_inputData_server(input, output, session, common_data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
