% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Report.R
\name{mod_Report_ui}
\alias{mod_Report_ui}
\alias{mod_Report_server}
\title{mod_Report_ui and mod_Report_server}
\usage{
mod_Report_ui(id)

mod_Report_server(
  input,
  output,
  session,
  common_data,
  detected_data,
  sample_metrics,
  inputData
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
