% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_plot_save.R
\name{multi_plot_save}
\alias{multi_plot_save}
\title{Save a ggplot object}
\usage{
multi_plot_save(file, plot, type, width, height, units, dpi)
}
\arguments{
\item{file}{Filepath to save plot to}

\item{plot}{A ggplot object}

\item{type}{One of "png", "pdf", "eps", or "tiff"}

\item{width}{Plot width}

\item{height}{Plot height}

\item{units}{One of "in", "cm", or "mm"}

\item{dpi}{Plot resolution in dpi}
}
\description{
Save a ggplot object
}
\examples{
\dontrun{ multi_plot_save("test.png",test_plot,"png",6,6,"in",600) }
}
\author{
Hugh Jones \email{hj.hujones.hugh@googlemail.com}
}
