% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_metrics}
\alias{sample_metrics}
\title{Collated sample metrics}
\format{
A list with 8 levels:
\describe{
\item{AIneg}{Data table holding collated negative mode AI mod sample metrics}
\item{AIpos}{Data table holding collated positive mode AI mod sample metrics}
\item{HCneg}{Data table holding collated negative mode H/C sample metrics}
\item{HCpos}{Data table holding collated positive mode H/C sample metrics}
\item{MWneg}{Data table holding collated negative mode MW sample metrics}
\item{MWpos}{Data table holding collated positive mode MW sample metrics}
\item{OCneg}{Data table holding collated negative mode O/C sample metrics}
\item{OCpos}{Data table holding collated positive mode O/C sample metrics}
}
Each data frame has 5 variables:
\describe{
\item{Blank}{Lab identifier}
\item{ESFA}{Value of metric for ESFA sample}
\item{PLFA}{Value of metric for PLFA sample} 
\item{SRFA}{Value of metric for SRFA sample}
\item{SRNOM}{Value of metric for SRNOM sample}
}
}
\usage{
sample_metrics
}
\description{
A list of 8 data frames
}
\seealso{
Other study data can be found in \code{common_data} and \code{detected_data}
}
\keyword{datasets}
