% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot2 theme as used by the Barrow group at Warwick}
\usage{
theme_custom(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{The base size, default is 11}

\item{base_family}{The base family, default is ""}

\item{base_line_size}{The base line size, default is \code{base_size}/22}

\item{base_rect_size}{The base rectangle size, default is \code{base_size}/22}
}
\value{
an ggplot2 theme
}
\description{
Custom ggplot2 theme as used by the Barrow group at Warwick
}
\author{
Hugh Jones \email{hj.hujones.hugh@googlemail.com}
}
